TypeScriptify Hera parser
