+++ source/hera.hera
@@ -1,3 +1,13 @@
+```
+import {
+  SequenceNode,
+  NameNode,
+  Handler,
+  StructuralTerminal,
+  StructuralObject
+} from './hera-types.civet'
+```
+
 Grammar
   Statement* ->
     const code = $1.filter(a => typeof a === "string")
@@ -25,17 +35,17 @@ Choice
   Sequence Handling ->
     if ($2 !== undefined) {
       if (!$1.push)
-        $1 = ["S", [$1], $2]
+        $1 = ["S", [$1], $2] as SequenceNode
       else
-        $1.push($2)
+        $1.push($2 as Handler)
     }
     return $1
 
 Sequence
-  Expression SequenceExpression+ ->
+  Expression SequenceExpression+ ::SequenceNode ->
     $2.unshift($1)
     return ["S", $2]
-  Expression ChoiceExpression+ ->
+  Expression ChoiceExpression+ ::SequenceNode ->
     $2.unshift($1)
     return ["/", $2]
   Expression
@@ -50,13 +60,13 @@ ParameterName
   ":" Name -> $2
 
 Expression
-  PrefixOperator? Suffix ParameterName? ->
+  PrefixOperator? Suffix ParameterName? ::SequenceNode|NameNode ->
     var result = null
     if ($1) result = [$1, $2]
     else result = $2
     if ($3)
-      return [{name: $3}, result]
-    return result
+      return [{name: $3}, result] as NameNode
+    return result as SequenceNode
 
 PrefixOperator
   [$&!]
@@ -82,7 +92,10 @@ Handling
   EOS ->
     return undefined
   Space* TypeAnnotation? Arrow HandlingExpression ->
-    if ($2) $4.t = $2
+    if ($2) {
+      if ($4 && typeof $4 === 'object' && 'f' in $4) $4.t = $2
+      else throw new Error(`TypeAnnotation is only supported for handler functions but you passed ${JSON.stringify($4)}`)
+    }
     return $4
 
 HandlingExpression
@@ -90,7 +103,7 @@ HandlingExpression
   StructuralMapping EOS -> $1
 
 HandlingExpressionBody
-  HandlingExpressionLine+ ->
+  HandlingExpressionLine+ ::Handler ->
     return {
       f: $1.join("").trimEnd(),
       $loc,
@@ -101,7 +114,7 @@ HandlingExpressionLine
 
 StructuralMapping
   StringValue ->
-    return JSON.parse(`"${$1}"`)
+    return JSON.parse(`"${$1}"`) as string
   NumberValue
   BooleanValue
   NullValue
@@ -110,13 +123,14 @@ StructuralMapping
   JSObject
 
 JSArray
-  OpenBracket ArrayItem* CloseBracket -> $2
+  OpenBracket ArrayItem* CloseBracket ::StructuralTerminal[] ->
+    return $2 as StructuralTerminal[]
 
 ArrayItem
   StructuralMapping /,\s*|\s*(?=\])/ -> $1
 
 JSObject
-  OpenBrace ObjectField* CloseBrace ->
+  OpenBrace ObjectField* CloseBrace ::StructuralObject ->
     return {
       o: Object.fromEntries($2)
     }
@@ -138,7 +152,7 @@ NullValue
   "null" ->
     return null
   "undefined" ->
-    return {l: undefined}
+    return { l: "undefined" }
 
 NumberValue
   /0x[\da-fA-F]+/ ->
